Moved all Components to specific directories
